+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ UNRELEASED
 ==========
 
 * ``cwd`` and ``environ`` are now defined on ``BaseFactory``
+* Add ``py.typed`` to state that the package is fully typed
 
 
 shell-utilities 1.0.0rc2 (2022-01-21)